+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [unreleased]
+### Fixed
+- bug where parse_url can return null for query and fragment components
+
+### Changed
+- removed unnecessary  PHPDocs
+- use anonymous class in test instead of additional class in the same file 
+
 ## [1.1.0] - 2019-03-08
 ### Added
 - CHANGELOG.md